+++ spec/support/shared_examples/gdal/major_object_examples.rb
@@ -49,23 +49,6 @@ RSpec.shared_examples 'a major object' do
     end
   end
 
-  describe '#description=' do
-    context 'new description is a string' do
-      around(:each) do |example|
-        old_description = subject.description
-
-        example.run
-
-        subject.description = old_description
-      end
-
-      it 'sets the items description' do
-        subject.description = 'a test description'
-        expect(subject.description).to eq 'a test description'
-      end
-    end
-  end
-
   describe '#null?' do
     it { is_expected.to_not be_null }
   end